+++ bmcd/src/authentication.rs
@@ -15,6 +15,7 @@ pub mod authentication_context;
 pub mod authentication_errors;
 pub mod authentication_service;
 pub mod ban_patrol;
+pub mod factory_password;
 pub mod linux_authenticator;
 pub mod passwd_validator;
 pub mod verified_proxy;